+++ Code/Gel/Collision/CollTriData.h
@@ -105,8 +105,8 @@ private:
 		CCollBSPNode *GetGreaterBranch() const { return IsLeftGreater() ? GetLeftBranch() : GetRightBranch(); }
 
 		// These two are needed for the cloning function
-		CCollBSPNode *GetBasePointer() const { return (CCollBSPNode*)(m_left_child_and_flags & ~0x3); }
-		void SetBasePointer(CCollBSPNode *p_base) { m_left_child_and_flags = ((uint32) p_base) | (m_left_child_and_flags & 0x3); }
+		CCollBSPNode *GetBasePointer() const { return (CCollBSPNode*)(m_left_child_and_flags & ~uintptr_t(0x3)); }
+		void SetBasePointer(CCollBSPNode *p_base) { m_left_child_and_flags = ((uintptr_t)p_base) | (m_left_child_and_flags & 0x3); }
 
 		bool IsLeftGreater() const { return m_left_child_and_flags & mLEFT_IS_GREATER; }
 		void SetLeftGreater(bool greater);
@@ -119,7 +119,7 @@ private:
 			mLEFT_IS_GREATER = 0x01, // Indicates that the left branch is the greater branch
 		};
 
-		uint32 m_left_child_and_flags = 0; // points to left branch, right branch one node over
+		uintptr_t m_left_child_and_flags = 0; // points to left branch, right branch one node over
 	};		
 
 protected:
@@ -136,7 +136,7 @@ protected:
 	struct SNode
 	{
 		int32				m_split_point = 0;		// the point on the axis (low 2 bits is the axis itself)
-		CCollBSPChildren	m_children = {};			// 32-bit value points to left branch, right branch one node over
+		CCollBSPChildren	m_children = {};		// 32-bit value points to left branch, right branch one node over
 	};
 
 	struct SLeaf
